Extract VaultCardView from VaultFragment.
